requirements.txt
disnake
requests
asyncio
python-dotenv
colorlog